Fix compile and apply code style.

+++ sql/400/01_characters.sql
@@ -12,8 +12,8 @@ DROP TABLE IF EXISTS `character_currency`;
 CREATE TABLE `character_currency` (
   `guid` INT(11) UNSIGNED NOT NULL,
   `currency` SMALLINT(5) UNSIGNED NOT NULL,
-  `count` SMALLINT(5) UNSIGNED NOT NULL,
-  `thisweek` SMALLINT(5) UNSIGNED NOT NULL,
+  `count` INT(11) UNSIGNED NOT NULL,
+  `thisweek` INT(11) UNSIGNED NOT NULL,
   PRIMARY KEY (`guid`,`currency`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
